Method to displays a welcome dialog box after saucerful login
This Java code defines a class called `WelcomeUser` within the `firstgame` package. The class contains a method named `showWelcomeDialog` that displays a welcome dialog box to the user for logging in or creating an account in a game app. The dialog box presents three options: "Create Account," "Custom Sign In," and "Sign in with Google."

Depending on the user's choice, the code takes different actions:

1. If the user selects "Create Account" (option 0), it creates an instance of the `CreateAcct` class, sets up the frame, displays it, and closes the initial welcome dialog box.

2. If the user selects "Custom Sign In" (option 1), the method returns, implying that the custom sign-in functionality is not implemented or not relevant in this section of the code.

3. If the user selects "Sign in with Google" (option 2), it creates an instance of the `LogInWithGoogle` class, initiates the frame, displays it, and closes the initial welcome dialog box.

This code seems to be a part of a graphical user interface (GUI) for a game app that provides options for users to create an account, sign in using custom credentials, or sign in with a Google account.